+++ app/controllers/admin/citations_controller.rb
@@ -2,8 +2,18 @@ class Admin::CitationsController < AdminController
   before_action :find_citation, except: :index
 
   def index
+    @query = params[:query]
+
+    citations = (
+      if @query
+        Citation.search(@query)
+      else
+        Citation
+      end
+    )
+
     @citations =
-      Citation.
+      citations.
       order(created_at: :desc).
       paginate(page: params[:page], per_page: 50)
   end